Polyglot instance accessible via react
